backend/requirements.txt
fastapi
uvicorn
pydantic
pytest
httpx
numpy
pandas
openai
faiss-cpu
